@@ -1,6 +1,6 @@
 module msmf
 
-go 1.15
+go 1.16
 
 require (
 	github.com/gorilla/mux v1.8.0
